Personal Part 3: Final Submission
Add a timer with setInterval to make the animation add new urban elements every 2.5 seconds to simulate the urbanization process. Set maxurbanizationStage to avoid infinite stacking of blocks in the animation. In addition, reset urbanizationStage to zero in windowresized() and regenerate urban elements to solve the previous problem that the animation could not run after window resize.